suppress compiler warning
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@10286 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ funcs/func_sha1.c
@@ -52,7 +52,7 @@ static int sha1(struct ast_channel *chan, char *cmd, char *data,
 	else {
 		ast_log(LOG_ERROR,
 				"Insufficient space to produce SHA1 hash result (%d < 41)\n",
-				len);
+				(int) len);
 	}
 
 	return 0;
